+++ python/lsst/ts/watcher/rules/__init__.py
@@ -30,7 +30,9 @@ from .hvac import *
 from .mt_air_compressors_state import *
 from .mt_ccw_following_rotator import *
 from .mt_force_error import *
+from .mt_m1m3_temperature import *
 from .mt_mirror_temperature import *
+from .mt_mount_azimuth import *
 from .mt_out_closed_loop_control import *
 from .mt_tangent_link_temperature import *
 from .mt_total_force_moment import *